+++ addons/main/display3DEN/menuStrip_tools_miscellaneous.hpp
@@ -35,7 +35,7 @@ class ENH_SwitchTime
 class ENH_ToggleGrass
 {
     text = "$STR_ENH_MAIN_TOOLS_TOGGLEGRASS";
-    action = "0 spawn ENH_fnc_toggleGrass";
+    action = "spawn ENH_fnc_toggleGrass";
     shortcuts[] = {INPUT_ALT_OFFSET + DIK_DOWN};
     wikiDescription = "Hides grass in Eden Editor.";
 };